Align README.md to library changes
I'm working through getting started with Qdrant and this library, and the code contained in the "Making Requests" section of the readme is not aligned with changes y'all have made in the last couple of weeks (VectorParams now has two additional fields, hnsw_config and quantization_config). The example in examples/search.rs does reflect these changes, but if someone is looking at the readme like I did, the example code won't work.